+++ src/main.c
@@ -21,6 +21,7 @@
 #include "nvic.h"
 #include "clocks.h"
 #include "adc.h"
+#include "rom.h"
 
 
 int any_core_running(void);
@@ -76,6 +77,7 @@ int main(int argc, char **argv) {
 
     /* Initialize CPU and peripherals */
     cpu_init();
+    rom_init();
     clocks_init();
     adc_init();